Add function to download all media from the thread
