+++ lib/pages/media_page.dart
@@ -1,13 +1,10 @@
 import 'dart:io';
-import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:flutter_chan/API/save_videos.dart';
 import 'package:flutter_chan/constants.dart';
 import 'package:flutter_chan/enums/enums.dart';
 import 'package:flutter_chan/widgets/webm_player.dart';
-import 'package:image_gallery_saver/image_gallery_saver.dart';
-import 'package:path_provider/path_provider.dart';
-import 'package:permission_handler/permission_handler.dart';
 import 'package:preload_page_view/preload_page_view.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
@@ -40,78 +37,10 @@ class MediaPage extends StatefulWidget {
 class _MediaPageState extends State<MediaPage> {
   MediaView mediaView = MediaView.pageView;
   PreloadPageController controller;
-  PageController pageController;
 
   final String page = '0';
   int index;
   String currentName = "";
-
-  Future<bool> _requestPermission(Permission permission) async {
-    if (await permission.isGranted) {
-      return true;
-    } else {
-      var result = await permission.request();
-      if (result == PermissionStatus.granted) {
-        return true;
-      }
-    }
-    return false;
-  }
-
-  Future<bool> saveVideo(
-      String url, String fileName, BuildContext context) async {
-    Directory directory;
-    var dio = Dio();
-
-    try {
-      if (Platform.isAndroid) {
-        if (await _requestPermission(Permission.storage)) {
-          directory = await getExternalStorageDirectory();
-          String newPath = "";
-          List<String> paths = directory.path.split("/");
-          for (int x = 1; x < paths.length; x++) {
-            String folder = paths[x];
-            if (folder != "Android") {
-              newPath += "/" + folder;
-            } else {
-              break;
-            }
-          }
-          newPath = newPath + "/Download/4Chan";
-          directory = Directory(newPath);
-        } else {
-          return false;
-        }
-      } else {
-        if (await _requestPermission(Permission.photos)) {
-          directory = await getTemporaryDirectory();
-        } else {
-          return false;
-        }
-      }
-
-      if (!await directory.exists()) {
-        await directory.create(recursive: true);
-      }
-      if (await directory.exists()) {
-        File saveFile = File(directory.path + "/$fileName");
-        await dio.download(url, saveFile.path,
-            onReceiveProgress: (value1, value2) {});
-        if (Platform.isIOS) {
-          await ImageGallerySaver.saveFile(saveFile.path,
-                  isReturnPathOfIOS: true)
-              .then((value) =>
-                  ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-                    content: Text('File downloaded!'),
-                  )));
-        }
-        return true;
-      }
-    } catch (e) {
-      print(e);
-    }
-    return false;
-  }
 
   onPageChanged(int i) {
     index = i;
@@ -132,11 +61,6 @@ class _MediaPageState extends State<MediaPage> {
     setStartVideo(widget.names[index]);
 
     controller = PreloadPageController(
-      initialPage: index,
-      keepPage: false,
-    );
-
-    pageController = PageController(
       initialPage: index,
       keepPage: false,
     );
@@ -173,10 +97,12 @@ class _MediaPageState extends State<MediaPage> {
             IconButton(
                 onPressed: () => {
                       saveVideo(
-                          'https://i.4cdn.org/${widget.board}/' +
-                              widget.fileNames[index],
-                          widget.video,
-                          context),
+                        'https://i.4cdn.org/${widget.board}/' +
+                            widget.fileNames[index],
+                        widget.video,
+                        context,
+                        true,
+                      ),
                     },
                 icon: Icon(Icons.download)),
           PopupMenuButton(
@@ -229,13 +155,6 @@ class _MediaPageState extends State<MediaPage> {
               onPageChanged: (i) => onPageChanged(i),
               preloadPagesCount: 2,
             )
-          // ? PageView(
-          //     scrollDirection: Axis.horizontal,
-          //     controller: pageController,
-          //     children: widget.list,
-          //     physics: ClampingScrollPhysics(),
-          //     onPageChanged: (i) => onPageChanged(i),
-          //   )
           : Column(
               children: [
                 Expanded(
