+++ lib/pages/thread_page.dart
@@ -3,6 +3,7 @@ import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_chan/API/api.dart';
+import 'package:flutter_chan/API/save_videos.dart';
 import 'package:flutter_chan/constants.dart';
 import 'package:flutter_chan/models/post.dart';
 import 'package:flutter_chan/widgets/floating_action_buttons.dart';
@@ -122,8 +123,6 @@ class _ThreadPageState extends State<ThreadPage> {
     });
   }
 
-  saveImages() {}
-
   @override
   void initState() {
     super.initState();
@@ -180,7 +179,11 @@ class _ThreadPageState extends State<ThreadPage> {
                   });
                 }
                 if (result == 1) {
-                  saveImages();
+                  saveAllMedia(
+                    'https://i.4cdn.org/${widget.board}/',
+                    fileNames,
+                    context,
+                  );
                 }
               })
         ],
